+++ com.ibm.streamsx.topology/opt/python/packages/streamsx/spl/spl.py
@@ -1104,6 +1104,19 @@ class for_each(object):
     Example SPL invocation::
 
         () as PT = PrintTuple(SensorReadings) { }
+        
+    Example definition with handling window punctuations::
+    
+        @spl.for_each(style='position')
+        class PrintPunct(object):
+            def __init__(self): 
+                pass
+
+            def __call__(self, value):
+                assert value > 0
+        
+            def on_punct(self):
+                print('window marker received')
 
     Args:
        style: How the SPL tuple is passed into Python callable, see  :ref:`spl-tuple-to-python`.
@@ -1114,6 +1127,9 @@ class for_each(object):
     If ``__exit__`` returns a true value when called with the exception 
     then the expression is suppressed and the tuple that caused the
     exception is ignored.
+    
+    Supports handling window punctuation markers in the Sink operator in ``on_punct`` method (new in version 1.16).
+    
     """
     def __init__(self, style=None, docpy=True):
         self.style = style
@@ -1159,6 +1175,19 @@ class PrimitiveOperator(object):
         """
         port_index = self._splpy_output_ports[port_id]
         ec._submit(self, port_index, tuple_)
+
+    def submit_punct(self, port_id):
+        """Submit a window punctuation marker to the output port.
+
+        Args:
+             port_id: Identifier of the port specified in the
+                  ``output_ports`` parameter of the ``@spl.primitive_operator``
+                  decorator.
+
+        .. versionadded:: 1.16
+        """
+        port_index = self._splpy_output_ports[port_id]
+        ec._submit_punct(self, port_index)
 
     def all_ports_ready(self):
         """Notifcation that the operator can submit tuples.
@@ -1302,6 +1331,21 @@ class primitive_operator(object):
                 near_match: 0.8;
         }
 
+    Example definition of an operator with punctuation handling::
+        
+        @spl.primitive_operator(output_ports=['A'])
+        class SimpleForwarder(spl.PrimitiveOperator):
+        def __init__(self):
+            pass
+
+        @spl.input_port()
+        def port0(self, *t):
+            self.submit('A', t)
+
+        def on_punct(self):
+            self.submit_punct('A')
+            
+    Supports handling window punctuation markers in the primitive operator in ``on_punct`` method (new in version 1.16).
 
     Args:
        output_ports(list): List of identifiers for output ports.