+++ test/python/spl/tests/test_splpy_primitives.py
@@ -177,6 +177,48 @@ class TestPrimitivesOutputs(unittest.TestCase):
         self.tester.contents(s, [{'v':9237}, {'v':-24}])
         self.tester.test(self.test_ctxtype, self.test_config)
 
+    def test_single_output_port_punct(self):
+        """Operator with single output port emits window marker before forwarding the tuple."""
+        topo = Topology()
+        streamsx.spl.toolkit.add_toolkit(topo, stu._tk_dir('testtkpy'))
+
+        s = topo.source([9237, -24])
+        s = s.map(lambda x : (x,), schema='tuple<int64 v>')
+
+        bop = op.Map("com.ibm.streamsx.topology.pytest.pyprimitives::SingleOutputPortPunct", s)
+
+        r = bop.stream
+        r.print(write_punctuations=True)
+    
+        self.tester = Tester(topo)
+        self.tester.tuple_count(r, 2)
+        self.tester.punct_count(r, 2)
+        self.tester.contents(s, [{'v':9237}, {'v':-24}])
+        self.tester.test(self.test_ctxtype, self.test_config)
+
+    def test_single_output_port_punct_forward(self):
+        """Operator with receives and forwards window marker."""
+        topo = Topology()
+        streamsx.spl.toolkit.add_toolkit(topo, stu._tk_dir('testtkpy'))
+
+        s = topo.source([1,2,3,4])
+        s = s.punctor(func=(lambda t : 2 < t), before=False)
+        s = s.map(lambda x : (x,), schema='tuple<int32 z>')
+
+        s1 = op.Map("com.ibm.streamsx.topology.pytest.pyprimitives::SingleOutputPort", s) # has no on_punct
+        bop = op.Map("com.ibm.streamsx.topology.pytest.pyprimitives::SingleOutputPortPunctForward", s) # implements on_punct
+
+        r = bop.stream
+        op.Sink("com.ibm.streamsx.topology.pytest.pyprimitives::VerifyPosInt", r)
+        
+        r.print(write_punctuations=True)
+        #self.test_config['topology.keepArtifacts'] = True
+    
+        self.tester = Tester(topo)
+        self.tester.tuple_count(r, 4)
+        self.tester.punct_count(r, 2)
+        self.tester.test(self.test_ctxtype, self.test_config)
+
     def test_multi_output_ports(self):
         """Operator with multiple output port."""
         topo = Topology()
@@ -191,6 +233,31 @@ class TestPrimitivesOutputs(unittest.TestCase):
     
         self.tester = Tester(topo)
         self.tester.tuple_count(s, 2)
+        self.tester.contents(r[0], [{'v1':9237}, {'v1':-24}])
+        self.tester.contents(r[1], [{'v2':9237+921}, {'v2':-24+921}])
+        self.tester.contents(r[2], [{'v3':9237-407}, {'v3':-24-407}])
+        self.tester.test(self.test_ctxtype, self.test_config)
+
+    def test_multi_output_ports_punct(self):
+        """Operator with multiple output port emits window marker after forwarding the tuple."""
+        topo = Topology()
+        streamsx.spl.toolkit.add_toolkit(topo, stu._tk_dir('testtkpy'))
+
+        s = topo.source([9237, -24])
+        s = s.map(lambda x : (x,), schema='tuple<int64 v>')
+
+        bop = op.Invoke(topo, "com.ibm.streamsx.topology.pytest.pyprimitives::MultiOutputPortsPunct", s, schemas=['tuple<int64 v1>', 'tuple<int32 v2>', 'tuple<int16 v3>'])
+
+        r = bop.outputs
+        r[0].print(tag='0', write_punctuations=True)
+        r[1].print(tag='1', write_punctuations=True)
+        r[2].print(tag='2', write_punctuations=True)
+    
+        self.tester = Tester(topo)
+        self.tester.tuple_count(s, 2)
+        self.tester.punct_count(r[0], 2)
+        self.tester.punct_count(r[1], 2)
+        self.tester.punct_count(r[2], 2)
         self.tester.contents(r[0], [{'v1':9237}, {'v1':-24}])
         self.tester.contents(r[1], [{'v2':9237+921}, {'v2':-24+921}])
         self.tester.contents(r[2], [{'v3':9237-407}, {'v3':-24-407}])